#1501 resolve issue for user login injection and cascading
