+++ src/FluentCMS.Web.UI/Components/Section.razor
@@ -5,11 +5,7 @@
     <div class="f-section f-section-@Name.FromPascalCaseToKebabCase()">
         @foreach (var Plugin in Page.Sections[Name].OrderBy(p => p.Order))
         {
-            <CascadingValue Value="@Plugin">
-                <PluginContainer @key="@Plugin">
-                    <DynamicComponent Type=@GetType(Plugin) />
-                </PluginContainer>
-            </CascadingValue>
+            <PluginContainer @key="@Plugin" Plugin=@Plugin SectionName=@Name />
         }
     </div>
 }
