+++ src/FluentCMS.Web.UI/Plugins/UserManagement/UserListPlugin.razor
@@ -1,7 +1,6 @@
 @inherits BasePlugin
-@rendermode InteractiveServer
 
-<PluginBody Title="@SectionName">
+<PluginBody Title="Users List">
     <ActionsFragment>
         <Button Href="@GetUrl("Create User")" Color="Color.Primary">
             <Icon Name="IconName.Plus" />Add User
